+++ __test__/index.dev.js
@@ -1,4 +1,0 @@
-// require all `test/components/**/index.js`
-const testsContext = require.context("./", true, /\.spec\.jsx?$/);
-
-testsContext.keys().forEach(testsContext);