+++ __test__/utils/Objects.spec.js
@@ -27,6 +27,45 @@ describe("utils/Objects.js", () => {
         };
         chai.assert.equal(Objects.sizeOf(object), 160);
     });
+    it("getTypeName", () => {
+
+    })
+    it("hasProperty", () => {
+        let obj = { element: "Example" };
+        chai.assert.isOk(Objects.hasProperty(obj, "element"));
+    })
+    it("mergeClone", () => {
+        let x = "X";
+        let y = "Y";
+        let z = "Z";
+        let a = { x, y };
+        let b = { y, z };
+        let result = Objects.mergeClone(a, b);
+        chai.assert.deepEqual(b, result);
+        chai.assert.isOk(b.x === "X");
+    })
+
+    it("cloneArray", () => {
+        let x = { obj: "example" };
+        let y = [x];
+        let result = Objects.cloneArray(y);
+        chai.assert.isOk(y !== result);
+    })
+
+    it("cloneObject", () => {
+        let object = {
+            bool: true,
+            number: 2,
+            string: "the string",
+            nestedObject: {
+                bool: false,
+                number: 3,
+                string: "another string"
+            }
+        };
+        chai.assert.deepEqual(Objects.clone(object), object);
+    })
+
     it("clone", () => {
         let object = {
             bool: true,
@@ -40,4 +79,6 @@ describe("utils/Objects.js", () => {
         };
         chai.assert.deepEqual(Objects.clone(object), object);
     });
+
+
 });