+++ config/server/JsonServer.js
@@ -1,12 +1,14 @@
+const Utility =  require("../util/Utility");
+const path = require("path");
 const jsonServer = require("json-server");
 
 const  config = {};
 
-config.createJsonServer = (port,routePath) => {
+config.createJsonServer = (port,routePath, done) => {
     // server.js
 
     const server = jsonServer.create();
-    const router = jsonServer.router(routePath);
+    const router = jsonServer.router(path.join(Utility.projectDir, routePath));
     const middlewares = jsonServer.defaults();
 
     server.use(middlewares);