+++ src/stores/Store.js
@@ -22,18 +22,13 @@ export default class Store extends BaseStore {
             this.onError = props.onError;
         }
         this.__dataMap = new MapArray([], this.__props.idField);
-
     }
 
     /**
      *
      * @param successCallBack
      * @param errorCallback
-     * @param offset
-     * @param limit
-     * @param query
-     * @param filter
-     * @param fields
+     * @param queryParams
      * @returns {boolean|*}
      */
     read(successCallBack: Function, errorCallback: Function, queryParams: Object): boolean {
@@ -105,7 +100,6 @@ export default class Store extends BaseStore {
     }
     /**
      *
-     * @param oldItem
      * @param newItem
      * @param successCallback
      * @param errorCallback